refactor(triggered/communicationmapping): move wizards to communicationmapping.ts
